pyproject.toml
[build-system]
requires = [
    "hatchling",
]
build-backend = "hatchling.build"

[project]
name = "universal-mcp-applications"
version = "0.1.21"
description = "A Universal MCP Application: universal_mcp_applications"
readme = "README.md"
requires-python = ">=3.11"
authors = [
    { name = "Manoj Bajaj", email = "manoj@agentr.dev" },
]
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "boto3>=1.38.32",
    "browser-use-sdk>=2.0.4",
    "dnspython>=2.4.0",
    "e2b-code-interpreter>=1.2.1",
    "elevenlabs>=2.14.0",
    "fal-client>=0.5.9",
    "firecrawl-py>=3.0.3",
    "google-genai>=1.33.0",
    "google-search-results>=2.4.2",
    "gql[all]>=3.5.2",
    "httpx-aiohttp>=0.1.8",
    "markitdown[audio-transcription,az-doc-intel,docx,outlook,pdf,pptx,xls,xlsx]>=0.1.3",
    "office365-rest-python-client>=2.6.2",
    "openai>=1.75.0",
    "pyairtable>=3.1.1",
    "pytest-asyncio>=1.1.0",
    "pytest>=8.3.5",
    "pytest>=8.4.1",
    "python-dotenv>=1.0.0",
    "replicate>=1.0.6",
    "requests>=2.31.0",
    "resend>=2.10.0",
    "twilio>=9.8.0",
    "universal-mcp>=0.1.24rc21",
    "yfinance>=0.2.66",
    "youtube-transcript-api==1.2.2",
]

[project.license]
text = "MIT"

[project.optional-dependencies]
test = [
    "pytest>=7.0.0,<9.0.0",
    "pytest-cov",
]
dev = [
    "ruff",
    "pre-commit",
]

[project.urls]
Homepage = "https://github.com/universal-mcp/applications"
Repository = "https://github.com/universal-mcp/applications"

[tool.hatch.envs.default]
dependencies = [
    "pytest",
    "pytest-cov",
]

[tool.hatch.envs.default.scripts]
test = "pytest {args:tests}"
test-cov = "pytest --cov-report term-missing --cov-config=pyproject.toml --cov=src/universal_mcp --cov=tests {args:tests}"
lint = "ruff check . && ruff format --check ."
format = "ruff format ."

[tool.hatch.build.targets.wheel]
packages = [
    "src/universal_mcp",
]

[tool.coverage.run]
source = [
    "src",
]
branch = true

[tool.coverage.report]
show_missing = true
fail_under = 70

[tool.ruff]
line-length = 140
lint.select = [
    "E", "W", "F", "I", "UP", "PL", "T20",
]
lint.ignore = []

[tool.ruff.format]
quote-style = "double"

[tool.pytest.ini_options]
pythonpath = [
    "src",
]

[dependency-groups]
dev = [
    "litellm>=1.77.3",
    "universal-mcp-agents>=0.1.13",
]
